requirements.txt
fastapi
diffusers
peft
wandb
Pillow
tqdm
accelerate 
git+https://github.com/huggingface/diffusers.git
opencv-python
lpips
google-cloud-storage
google-auth
fvcore
omegaconf
av
pyngrok
cloudpickle
pycocotools
python-multipart
dagshub
mlflow
pydantic